New functionalities

+++ src/main/java/pl/projectarea/project0/pricealert/PriceAlertRepository.java
@@ -8,12 +8,12 @@ import pl.projectarea.project0.pricealert.PriceAlert;
 import java.util.List;
 
 @Repository
-public interface PriceAlertRepository extends JpaRepository<PriceAlert, Integer> {
+public interface PriceAlertRepository extends JpaRepository<PriceAlert, Long> {
 
     @Query(nativeQuery = true, value = "SELECT FROM price_alert WHERE ticker=?1")
     List<PriceAlert> findAllByTicker(String ticker);
 
     @Query(nativeQuery = true, value = "SELECT FROM price_alert WHERE user_id=?1")
-    List<PriceAlert> findAllByUserId(Integer id);
+    List<PriceAlert> findAllByUserId(Long id);
 
 }
